+++ packages/utils/ts-utils/src/array/slice/safe-slice.test.ts
@@ -1,3 +1,4 @@
+import { uint32 } from '../../types';
 import { testArrayEquality } from '../array-tester';
 import { safeSlice } from './safe-slice';
 
@@ -5,56 +6,56 @@ const target = [0, 1, 2, 3, 4];
 
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, 0, 5),
+  target: safeSlice(target, 0 as uint32, 5 as uint32),
   toBe: target,
 }); // 正常
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, 0, 6),
+  target: safeSlice(target, 0 as uint32, 6 as uint32),
   toBe: target,
 }); // 片方オーバー
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, -1, 5),
+  target: safeSlice(target, -1 as uint32, 5 as uint32),
   toBe: target,
 }); // 片方オーバー
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, -1, 6),
+  target: safeSlice(target, -1 as uint32, 6 as uint32),
   toBe: target,
 }); // 両方オーバー
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, 0, 3),
+  target: safeSlice(target, 0 as uint32, 3 as uint32),
   toBe: [0, 1, 2],
 }); // 正常
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, -1, 3),
+  target: safeSlice(target, -1 as uint32, 3 as uint32),
   toBe: [0, 1, 2],
 }); // 片方オーバー
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, 3, 5),
+  target: safeSlice(target, 3 as uint32, 5 as uint32),
   toBe: [3, 4],
 }); // 正常
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, 3, 6),
+  target: safeSlice(target, 3 as uint32, 6 as uint32),
   toBe: [3, 4],
 }); // 片方オーバー
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, 4, 3),
+  target: safeSlice(target, 4 as uint32, 3 as uint32),
   toBe: [],
 }); // start > end
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, 0, -1),
+  target: safeSlice(target, 0 as uint32, -1 as uint32),
   toBe: [],
 }); // start > end
 testArrayEquality({
   testName: 'safeSlice',
-  target: safeSlice(target, -1, -2),
+  target: safeSlice(target, -1 as uint32, -2 as uint32),
   toBe: [],
 }); // start > end
